+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.25.0] - 2025-03-19
+
 ### Changed
 
 - resource-definition-list: add more state values CLONING/FAILED...